+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v1.0.1
+
+* Adding default values to help text
+* Print arguments if verbose set
+
 ## v1.0.0
 
 * Initial release